# Add ASM optimizations for MuHash3072