Adds assembly optimizations for `MuHash3072` which is used for the `muhash` calculation of the UTXO set hash.