@Sjors 2020-06-09
I'll post some results here after #19214 and when the SHA512 -> 256 [switch](https://github.com/bitcoin/bitcoin/pull/19055#issuecomment-640875012) is implemented.