@fjahr 2021-12-19
Thanks for the nudge @Sjors and @DrahtBot !

I have only done a plain rebase so far, i.e. I have not checked whether further optimizations may be interesting to look into. But while I ran the benchmarks again I noticed that it currently appears that this has slower benchmarks than master on my machine. `mul` and `div` operations appear to be almost 20% slower. I will try to test on some other machines soon but if someone else can give it a spin as well it would be a great help!

<details><summary>Results with this PR</summary>
<p>
```console
$ src/bench/bench_bitcoin -filter=MuHash.* -min_time=1000

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            7,184.06 |          139,197.00 |    0.8% |      1.07 | `MuHash`
|            6,138.95 |          162,894.31 |    0.1% |      1.05 | `MuHashDiv`
|            6,161.80 |          162,290.16 |    0.7% |      1.05 | `MuHashMul`
|            1,016.96 |          983,326.67 |    0.3% |      1.08 | `MuHashPrecompute`
```
</p>
</details>

<details><summary>Results on master</summary>
<p>
```console
$ src/bench/bench_bitcoin -filter=MuHash.* -min_time=1000

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            6,225.79 |          160,622.20 |    1.6% |      1.05 | `MuHash`
|            5,121.86 |          195,241.46 |    0.5% |      1.09 | `MuHashDiv`
|            5,173.32 |          193,299.36 |    1.5% |      1.11 | `MuHashMul`
|            1,023.41 |          977,127.42 |    0.7% |      1.08 | `MuHashPrecompute`
```
</p>
</details>