@sipa 2021-12-20
@MarcoFalke Well, that's expected at least. There are only x86_64 asm implementations here.

@fjahr What kind of system/compiler/flags?