@sipa 2021-12-20
Benchmark on Ryzen 5950X:

master:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            4,133.94 |          241,900.24 |    0.5% |      0.01 | `MuHash`
|            3,530.11 |          283,277.63 |    0.4% |      0.01 | `MuHashDiv`
|            3,522.44 |          283,894.43 |    0.2% |      0.01 | `MuHashMul`
|              554.91 |        1,802,107.77 |    0.1% |      0.01 | `MuHashPrecompute`

This PR:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            3,157.98 |          316,658.14 |    0.3% |      0.01 | `MuHash`
|            2,597.48 |          384,987.98 |    0.1% |      0.01 | `MuHashDiv`
|            2,597.10 |          385,044.86 |    0.2% |      0.01 | `MuHashMul`
|              556.71 |        1,796,271.37 |    0.0% |      0.01 | `MuHashPrecompute`