@MarcoFalke 2021-12-21
> @MarcoFalke Well, that's expected at least. There are only x86_64 asm implementations here.

I wish I could read asm. :sweat_smile: