@fjahr 2021-12-21
> @fjahr What kind of system/compiler/flags?

Hmm, this was with Intel Core i5-6287U, darwin 21.2.0 (macOS 12.1), clang 13.0.0, default configure flags except for skipping the GUI.