@sipa 2021-12-21
@MarcoFalke 

> I wish I could read asm. :sweat_smile: 

The `#if defined(__amd64__) || defined(__x86_64__)` also helps.