@fjahr 2022-11-13
> This seems to be just a matter of single unadressed review comment and more benchmarking. I could tak over if @fjahr is not interested anymore.

Thanks, but I am still happy to keep this alive. I will try to run benchmarks again myself, I am on a new machine yet again now. Of course more benchmarks help, so if you could run them as well that would be great @kristapsk .

I addressed the comment from @fanquake and rebased the branch.

@achow101 please reopen. I was a bit busy and forgot to respond here but I will try to not let it happen in the future :)