@fjahr 2022-11-14
> Since you've force pushed to the branch, GitHub won't allow this to be reopened. I think if you push the commit that was the HEAD at the time of closing, it can be reopened. Or you can make a new PR.

Alright, I pushed the old commit back to the remote branch.