@real-or-random 2023-04-25
> Based on those results, if at all, those optimizations should probably only be enabled if GCC is used? (Not sure if it's worth the review and maintenance burden though.)

Some of the code changes here conflict with https://github.com/bitcoin/bitcoin/pull/21590, and given that clang 14 is better than this ASM, we should maybe get benchmarks on a recent GCC (like 12.2) and see if it's faster. If not, I also wonder if there's a better way to convince GCC to output good code.

 @theStack Can you re-benchmark with GCC 12 ? I can also try to get some numbers on my machine.