@fjahr 2023-05-01
I finally got another amd64 machine with which I can test again. I confirmed the clang-14 results others have seen and I tested with GCC 13.1. The results still look good there for me but would be great if someone else could confirm.

(using `src/bench/bench_bitcoin -filter=MuHash.* -min-time=1000`)

Master:
```
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            9,000.29 |          111,107.54 |    1.6% |      1.09 | `MuHash`
|            7,424.35 |          134,691.90 |    0.0% |      1.05 | `MuHashDiv`
|            7,428.86 |          134,610.13 |    0.2% |      1.06 | `MuHashMul`
|            1,058.97 |          944,314.00 |    0.0% |      1.08 | `MuHashPrecompute`
```

This PR:
```
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            6,936.64 |          144,161.92 |    0.1% |      1.08 | `MuHash`
|            5,596.30 |          178,689.37 |    0.2% |      1.03 | `MuHashDiv`
|            5,422.52 |          184,415.94 |    0.7% |      1.08 | `MuHashMul`
|            1,061.63 |          941,950.39 |    0.0% |      1.09 | `MuHashPrecompute`
```

I guess it makes sense to only use this when GCC is used which should work with this 
```
#if (defined(__amd64__) || defined(__x86_64__)) && defined(__GNUC__) && !defined(__clang__)
```
but I would also be curious if there is another way to tell GCC to do the same optimizations clang seems to be doing.

I will also try to combine the changes here with #21590 to see what the combined impact is.